Factor out handle_remote_data to libqrexec
Changes:
- introduce constants for return codes
- extract options for replacing characters and calling shutdown
